// packages/rpc/src/types/api.d.ts
 declare namespace CKBComponents {
  export type DAO = string
  export type Hash = string
  export type Number = string
  export type Hash256 = string
  export type UInt32 = number
  export type Index = string
  export type Version = string
  export type Count = string
  export type Difficulty = string
  export type BlockNumber = string
  export type EpochInHeader = string
  export type Capacity = string
  export type ProposalShortId = string
  export type Timestamp = string
  export type Nonce = string
  export type Cycles = string
  export type Size = string
  export type OutputsValidator = 'default' | 'passthrough' | undefined
  export type RationalU256 = Record<'denom' | 'numer', string>
  export type ProposalWindow = Record<'closest' | 'farthest', BlockNumber>
  export type EpochNumberWithFraction = string
  export enum TransactionStatus {
    Pending = 'pending',
    Proposed = 'proposed',
    Committed = 'committed',
  }

  export type ScriptHashType = 'data' | 'type' | 'data1'

  export type DepType = 'code' | 'depGroup'
  export type JsonBytes = string

  /**
   * @typedef Bytes, keep consistent with CKB
   * @description Bytes will be serialized to string
   * @see https://github.com/nervosnetwork/ckb/blob/develop/util/jsonrpc-types/src/blockchain.rs#L19
   */
  export type Bytes = string
  export type Since = string
// ...
  export interface Script {
    args: Bytes
    codeHash: Hash256
    hashType: ScriptHashType
  }

  /**
   * @typedef CellInput, cell input in a transaction
   * @property previousOutput, point to its P1 cell
   * @property since, a parameter to prevent a cell to be spent before a centain block timestamp or a block number,
   *           [RFC](https://github.com/nervosnetwork/rfcs/blob/master/rfcs/0017-tx-valid-since/0017-tx-valid-since.md)
   */
  export interface CellInput {
    previousOutput: OutPoint | null
    since: Since
  }

  /**
   * @typedef CellOutput, cell output in a transaction
   * @property capacity, the capacity of the genereated P1 cell
   * @property lock, lock script
   * @property type, type script
   */
  export interface CellOutput {
    capacity: Capacity
    lock: Script
    type?: Script | null
  }

  /**
   * @typedef OutPoint, used to refer a generated cell by transaction hash and output index
   * @property hash, transaction hash
   * @property index, index of cell output
   */
  export interface OutPoint {
    txHash: Hash256
    index: Index
  }

  /**
   * @typeof CellDep, cell dependencies in a transaction
   * @property outPoint, the out point of the cell dependency
   * @property depType, indicate if the data of the cell containing a group of dependencies
   */
  export interface CellDep {
    outPoint: OutPoint | null
    depType: DepType
  }

  export type Witness = Bytes
// ...
  export interface RawTransaction {
    version: Version
    cellDeps: CellDep[]
    headerDeps: Hash256[]
    inputs: CellInput[]
    outputs: CellOutput[]
    witnesses: Witness[]
    outputsData: Bytes[]
  }
// ...
  export interface TransactionProof {
    blockHash: Hash
    proof: {
      indices: Number[]
      lemmas: Hash[]
    }
    witnessesRoot: Hash
  }
// ...
}